Refactor cyral_sidecar_listener
