+++ cyral/internal/sidecar/listener/constants.go
@@ -1,0 +1,24 @@
+package listener
+
+// Resource
+const (
+	resourceName   = "cyral_sidecar_listener"
+	dataSourceName = "cyral_sidecar_listener"
+
+	RepoTypesKey         = "repo_types"
+	NetworkAddressKey    = "network_address"
+	MySQLSettingsKey     = "mysql_settings"
+	DbVersionKey         = "db_version"
+	CharacterSetKey      = "character_set"
+	S3SettingsKey        = "s3_settings"
+	ProxyModeKey         = "proxy_mode"
+	DynamoDbSettingsKey  = "dynamodb_settings"
+	SQLServerSettingsKey = "sqlserver_settings"
+	SQLServerVersionKey  = "version"
+)
+
+// Data source
+const (
+	SidecarListenerListKey = "listener_list"
+	DSRepoTypeKey          = "repo_type"
+)
