+++ cyral/internal/sidecar/listener/resource.go
@@ -3,228 +3,34 @@ package listener
 import (
 	"context"
 	"fmt"
-	"net/http"
 
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 	"github.com/cyralinc/terraform-provider-cyral/cyral/client"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core"
-	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/operationtype"
+	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/resourcetype"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/utils"
 )
 
-// create a constant block for schema keys
-
-const (
-	RepoTypesKey         = "repo_types"
-	NetworkAddressKey    = "network_address"
-	MySQLSettingsKey     = "mysql_settings"
-	DbVersionKey         = "db_version"
-	CharacterSetKey      = "character_set"
-	S3SettingsKey        = "s3_settings"
-	ProxyModeKey         = "proxy_mode"
-	DynamoDbSettingsKey  = "dynamodb_settings"
-	SQLServerSettingsKey = "sqlserver_settings"
-	SQLServerVersionKey  = "version"
-)
-
-// SidecarListener struct for sidecar listener.
-type SidecarListener struct {
-	SidecarId         string             `json:"-"`
-	ListenerId        string             `json:"id"`
-	RepoTypes         []string           `json:"repoTypes"`
-	NetworkAddress    *NetworkAddress    `json:"address,omitempty"`
-	MySQLSettings     *MySQLSettings     `json:"mysqlSettings,omitempty"`
-	S3Settings        *S3Settings        `json:"s3Settings,omitempty"`
-	DynamoDbSettings  *DynamoDbSettings  `json:"dynamoDbSettings,omitempty"`
-	SQLServerSettings *SQLServerSettings `json:"sqlServerSettings,omitempty"`
-}
-type NetworkAddress struct {
-	Host string `json:"host,omitempty"`
-	Port int    `json:"port"`
-}
-type MySQLSettings struct {
-	DbVersion    string `json:"dbVersion,omitempty"`
-	CharacterSet string `json:"characterSet,omitempty"`
-}
-type S3Settings struct {
-	ProxyMode bool `json:"proxyMode,omitempty"`
-}
-type DynamoDbSettings struct {
-	ProxyMode bool `json:"proxyMode,omitempty"`
-}
-
-type SQLServerSettings struct {
-	Version string `json:"version,omitempty"`
-}
-
-var ReadSidecarListenersConfig = core.ResourceOperationConfig{
-	ResourceName: "SidecarListenersResourceRead",
-	Type:         operationtype.Read,
-	HttpMethod:   http.MethodGet,
-	URLFactory: func(d *schema.ResourceData, c *client.Client) string {
+var resourceContextHandler = core.DefaultContextHandler{
+	ResourceName:                  resourceName,
+	ResourceType:                  resourcetype.Resource,
+	SchemaReaderFactory:           func() core.SchemaReader { return &SidecarListenerResource{} },
+	SchemaWriterFactoryGetMethod:  func(_ *schema.ResourceData) core.SchemaWriter { return &ReadSidecarListenerAPIResponse{} },
+	SchemaWriterFactoryPostMethod: func(_ *schema.ResourceData) core.SchemaWriter { return &CreateListenerAPIResponse{} },
+	BaseURLFactory: func(d *schema.ResourceData, c *client.Client) string {
+		return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners",
+			c.ControlPlane,
+			d.Get(utils.SidecarIDKey).(string))
+	},
+	IdBasedURLFactory: func(d *schema.ResourceData, c *client.Client) string {
 		return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners/%s",
 			c.ControlPlane,
 			d.Get(utils.SidecarIDKey).(string),
-			d.Get(utils.ListenerIDKey).(string))
+			d.Get(utils.ListenerIDKey).(string),
+		)
 	},
-	SchemaWriterFactory: func(_ *schema.ResourceData) core.SchemaWriter { return &ReadSidecarListenerAPIResponse{} },
-	RequestErrorHandler: &core.IgnoreHttpNotFound{ResName: "Sidecar listener"},
-}
-
-type ReadSidecarListenerAPIResponse struct {
-	ListenerConfig *SidecarListener `json:"listenerConfig"`
-}
-type CreateListenerAPIResponse struct {
-	ListenerId string `json:"listenerId"`
-}
-
-func (c CreateListenerAPIResponse) WriteToSchema(d *schema.ResourceData) error {
-	d.SetId(utils.MarshalComposedID([]string{d.Get(utils.SidecarIDKey).(string), c.ListenerId}, "/"))
-	return d.Set(utils.ListenerIDKey, c.ListenerId)
-}
-
-func (data ReadSidecarListenerAPIResponse) WriteToSchema(d *schema.ResourceData) error {
-	ctx := context.Background()
-	tflog.Debug(ctx, "Init ReadSidecarListenerAPIResponse.WriteToSchema")
-	if data.ListenerConfig != nil {
-		_ = d.Set(utils.ListenerIDKey, data.ListenerConfig.ListenerId)
-		_ = d.Set(RepoTypesKey, data.ListenerConfig.RepoTypesAsInterface())
-		_ = d.Set(NetworkAddressKey, data.ListenerConfig.NetworkAddressAsInterface())
-		_ = d.Set(S3SettingsKey, data.ListenerConfig.S3SettingsAsInterface())
-		_ = d.Set(MySQLSettingsKey, data.ListenerConfig.MySQLSettingsAsInterface())
-		_ = d.Set(DynamoDbSettingsKey, data.ListenerConfig.DynamoDbSettingsAsInterface())
-		_ = d.Set(SQLServerSettingsKey, data.ListenerConfig.SQLServerSettingsAsInterface())
-	}
-	tflog.Debug(ctx, "End ReadSidecarListenerAPIResponse.WriteToSchema")
-	return nil
-}
-
-func (l *SidecarListener) RepoTypesAsInterface() []interface{} {
-	if l.RepoTypes == nil {
-		return nil
-	}
-	result := make([]interface{}, len(l.RepoTypes))
-	for i, v := range l.RepoTypes {
-		result[i] = v
-	}
-	return result
-}
-func (l *SidecarListener) RepoTypesFromInterface(anInterface []interface{}) {
-	repoTypes := make([]string, len(anInterface))
-	for i, v := range anInterface {
-		repoTypes[i] = v.(string)
-	}
-	l.RepoTypes = repoTypes
-}
-func (l *SidecarListener) NetworkAddressAsInterface() []interface{} {
-	if l.NetworkAddress == nil {
-		return nil
-	}
-	return []interface{}{
-		map[string]interface{}{
-			utils.HostKey: l.NetworkAddress.Host,
-			utils.PortKey: l.NetworkAddress.Port,
-		},
-	}
-}
-func (l *SidecarListener) NetworkAddressFromInterface(anInterface []interface{}) {
-	if len(anInterface) == 0 {
-		return
-	}
-	l.NetworkAddress = &NetworkAddress{
-		Host: anInterface[0].(map[string]interface{})[utils.HostKey].(string),
-		Port: anInterface[0].(map[string]interface{})[utils.PortKey].(int),
-	}
-}
-func (l *SidecarListener) MySQLSettingsAsInterface() []interface{} {
-	if l.MySQLSettings == nil {
-		return nil
-	}
-	return []interface{}{map[string]interface{}{
-		DbVersionKey:    l.MySQLSettings.DbVersion,
-		CharacterSetKey: l.MySQLSettings.CharacterSet,
-	}}
-}
-func (l *SidecarListener) MySQLSettingsFromInterface(anInterface []interface{}) {
-	if len(anInterface) == 0 {
-		return
-	}
-	l.MySQLSettings = &MySQLSettings{
-		DbVersion:    anInterface[0].(map[string]interface{})[DbVersionKey].(string),
-		CharacterSet: anInterface[0].(map[string]interface{})[CharacterSetKey].(string),
-	}
-}
-func (l *SidecarListener) S3SettingsAsInterface() []interface{} {
-	if l.S3Settings == nil {
-		return nil
-	}
-	return []interface{}{map[string]interface{}{
-		ProxyModeKey: l.S3Settings.ProxyMode,
-	}}
-}
-func (l *SidecarListener) S3SettingsFromInterface(anInterface []interface{}) {
-	if len(anInterface) == 0 {
-		return
-	}
-	l.S3Settings = &S3Settings{
-		ProxyMode: anInterface[0].(map[string]interface{})[ProxyModeKey].(bool),
-	}
-}
-func (l *SidecarListener) DynamoDbSettingsAsInterface() []interface{} {
-	if l.DynamoDbSettings == nil {
-		return nil
-	}
-	return []interface{}{map[string]interface{}{
-		ProxyModeKey: l.DynamoDbSettings.ProxyMode,
-	}}
-}
-func (l *SidecarListener) DynamoDbSettingsFromInterface(anInterface []interface{}) {
-	if len(anInterface) == 0 {
-		return
-	}
-	l.DynamoDbSettings = &DynamoDbSettings{
-		ProxyMode: anInterface[0].(map[string]interface{})[ProxyModeKey].(bool),
-	}
-}
-func (l *SidecarListener) SQLServerSettingsAsInterface() []interface{} {
-	if l.SQLServerSettings == nil {
-		return nil
-	}
-	return []interface{}{map[string]interface{}{
-		SQLServerVersionKey: l.SQLServerSettings.Version,
-	}}
-}
-func (l *SidecarListener) SQLServerSettingsFromInterface(anInterface []interface{}) {
-	if len(anInterface) == 0 {
-		return
-	}
-	l.SQLServerSettings = &SQLServerSettings{
-		Version: anInterface[0].(map[string]interface{})[SQLServerVersionKey].(string),
-	}
-}
-
-// SidecarListenerResource represents the payload of a create or update a listener request
-type SidecarListenerResource struct {
-	ListenerConfig SidecarListener `json:"listenerConfig"`
-}
-
-// ReadFromSchema populates the SidecarListenerResource from the schema
-func (s *SidecarListenerResource) ReadFromSchema(d *schema.ResourceData) error {
-	s.ListenerConfig = SidecarListener{
-		SidecarId:  d.Get(utils.SidecarIDKey).(string),
-		ListenerId: d.Get(utils.ListenerIDKey).(string),
-	}
-	s.ListenerConfig.RepoTypesFromInterface(d.Get(RepoTypesKey).([]interface{}))
-	s.ListenerConfig.NetworkAddressFromInterface(d.Get(NetworkAddressKey).(*schema.Set).List())
-	s.ListenerConfig.MySQLSettingsFromInterface(d.Get(MySQLSettingsKey).(*schema.Set).List())
-	s.ListenerConfig.S3SettingsFromInterface(d.Get(S3SettingsKey).(*schema.Set).List())
-	s.ListenerConfig.DynamoDbSettingsFromInterface(d.Get(DynamoDbSettingsKey).(*schema.Set).List())
-	s.ListenerConfig.SQLServerSettingsFromInterface(d.Get(SQLServerSettingsKey).(*schema.Set).List())
-
-	return nil
 }
 
 // resourceSidecarListener returns the schema and methods for provisioning a sidecar listener
@@ -233,55 +39,16 @@ func (s *SidecarListenerResource) ReadFromSchema(d *schema.ResourceData) error {
 // POST {{baseURL}}/sidecars/:sidecarID/listeners/ (Create a listener)
 // PUT {{baseURL}}/sidecars/:sidecarID/listeners/:listenerID (Update a listener)
 // DELETE {{baseURL}}/sidecars/:sidecarID/listeners/:listenerID (Delete a listener)
-func ResourceSidecarListener() *schema.Resource {
+func resourceSchema() *schema.Resource {
 	return &schema.Resource{
 		Description: "Manages [sidecar listeners](https://cyral.com/docs/sidecars/sidecar-listeners)." +
 			"\n~> **Warning** Multiple listeners can be associated to a single sidecar as long as " +
 			"`host` and `port` are unique. If `host` is omitted, then `port` must be unique.",
-		CreateContext: core.CreateResource(
-			core.ResourceOperationConfig{
-				ResourceName: "SidecarListenersResourceCreate",
-				Type:         operationtype.Create,
-				HttpMethod:   http.MethodPost,
-				URLFactory: func(d *schema.ResourceData, c *client.Client) string {
-					return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners",
-						c.ControlPlane,
-						d.Get(utils.SidecarIDKey).(string))
+		CreateContext: resourceContextHandler.CreateContext(),
+		ReadContext:   resourceContextHandler.ReadContext(),
+		UpdateContext: resourceContextHandler.UpdateContext(),
+		DeleteContext: resourceContextHandler.DeleteContext(),
 
-				},
-				SchemaReaderFactory: func() core.SchemaReader { return &SidecarListenerResource{} },
-				SchemaWriterFactory: func(_ *schema.ResourceData) core.SchemaWriter { return &CreateListenerAPIResponse{} },
-			}, ReadSidecarListenersConfig,
-		),
-		ReadContext: core.ReadResource(ReadSidecarListenersConfig),
-		UpdateContext: core.UpdateResource(
-			core.ResourceOperationConfig{
-				ResourceName: "SidecarListenersResourceUpdate",
-				Type:         operationtype.Update,
-				HttpMethod:   http.MethodPut,
-				URLFactory: func(d *schema.ResourceData, c *client.Client) string {
-					return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners/%s",
-						c.ControlPlane,
-						d.Get(utils.SidecarIDKey).(string),
-						d.Get(utils.ListenerIDKey).(string))
-
-				},
-				SchemaReaderFactory: func() core.SchemaReader { return &SidecarListenerResource{} },
-			}, ReadSidecarListenersConfig,
-		),
-		DeleteContext: core.DeleteResource(
-			core.ResourceOperationConfig{
-				ResourceName: "SidecarListenersResourceDelete",
-				Type:         operationtype.Delete,
-				HttpMethod:   http.MethodDelete,
-				URLFactory: func(d *schema.ResourceData, c *client.Client) string {
-					return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners/%s",
-						c.ControlPlane,
-						d.Get(utils.SidecarIDKey).(string),
-						d.Get(utils.ListenerIDKey).(string))
-				},
-			},
-		),
 		Schema: getSidecarListenerSchema(),
 		Importer: &schema.ResourceImporter{
 			StateContext: func(
