+++ cyral/internal/sidecar/listener/schema_loader.go
@@ -1,4 +1,4 @@
-package sidecar
+package listener
 
 import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core"
@@ -13,6 +13,11 @@ func (p *packageSchema) Name() string {
 
 func (p *packageSchema) Schemas() []*core.SchemaDescriptor {
 	return []*core.SchemaDescriptor{
+		{
+			Name:   dataSourceName,
+			Type:   core.DataSourceSchemaType,
+			Schema: dataSourceSchema,
+		},
 		{
 			Name:   resourceName,
 			Type:   core.ResourceSchemaType,
